mofa/requirements.txt
Django==2.2.6
djangorestframework==3.12.1
django-nose==1.4.7
python-dotenv==0.10.3
pytz==2019.2
sqlparse==0.3.0
requests==2.22.0
flake8==3.7.8
apscheduler==3.6.1
coverage==4.5.4
psycopg2-binary==2.8.6
flake8-docstrings==1.5.0
sphinx==2.2.1
sphinx-autoapi==1.2.1
sphinx-rtd-theme==0.4.3
sphinx-fontawesome==0.0.6
djangorestframework==3.12.1
numpy==1.19.5
pandas==1.2.1
matplotlib==3.3.3
scikit-learn==0.24.1
git+https://github.com/digi604/django-smart-selects.git@js-unlinting-fixes
